Add backward tests, split test files
